chore(login):add user service for login
[Delivers #167750530]
