+++ app/controller/UsersController.js
@@ -1,5 +1,3 @@
-import createToken from '../../lib/helpers/jwtToken'
-import config from '../../config/config'
 import UserService from '../services/UserService'
 
 class UserController {
@@ -12,14 +10,7 @@ class UserController {
   }
 
   static async loginUser(req, res) {
-    const token = createToken(
-      { id: req.user.id, role: req.user.role },
-      config.secretKey,
-      { expiresIn: config.jwtExpiration },
-    )
-    return res
-      .status(200)
-      .send({ status: 'success', message: 'LoggedIn successfully', token })
+    UserService.login(req, res)
   }
 }
