+++ app/services/UserService.js
@@ -1,6 +1,8 @@
 import UserModel from '../models/User'
 import AdminModel from '../models/Admin'
 import EncryptData from '../../lib/helpers/EncryptPassword'
+import createToken from '../../lib/helpers/jwtToken'
+import config from '../../config/config'
 
 class UserService {
   static async registerUser(req, res) {
@@ -23,6 +25,17 @@ class UserService {
     req.body.password = EncryptData.generateHash(req.body.password)
     const message = await AdminModel.createAdmin(req.body)
     return res.status(201).send({ status: 201, message })
+  }
+
+  static async login(req, res) {
+    const token = createToken(
+      { id: req.user.id, role: req.user.role },
+      config.secretKey,
+      { expiresIn: config.jwtExpiration },
+    )
+    return res
+      .status(200)
+      .send({ status: 200, message: 'Logged In successfully', token })
   }
 }
